refactor: move htmlReport into its own action/reducer files
